requirements/requirements-dev.txt
django-extensions==3.2.0
ipython==8.4.0
mycli==1.25.0
bcrypt==3.2.2
pyjwt==2.4.0
